+++ rspace-bench/src/main/scala/coop/rchain/rspace/bench/EvalBenchStateBase.scala
@@ -24,7 +24,7 @@ trait EvalBenchStateBase {
 
   @Setup
   def doSetup(): Unit = {
-    EvalBenchStateBase.deleteOldStorage(dbDir)
+    deleteOldStorage(dbDir)
 
     term = Interpreter.buildNormalizedTerm(resourceFileReader(rhoScriptSource)).runAttempt match {
       case Right(par) => Some(par)
@@ -44,25 +44,4 @@ trait EvalBenchStateBase {
       Option(getClass.getResourceAsStream(path))
         .getOrElse(throw new FileNotFoundException(path))
     )
-}
-
-object EvalBenchStateBase {
-
-  /**
-    * until we add flag 'delete_lmdb_dir_on_close' for benchmarks and unit-tests
-    * this prevents periodic out of disk space failures
-    */
-  def deleteOldStorage(dbDir: Path): Unit =
-    dbDir.getParent.toFile.listFiles
-      .filter(dir => dir.isDirectory && (dir.toPath != dbDir))
-      .filter(_.getName.startsWith("rchain-storage-test-"))
-      .foreach(
-        dir =>
-          try {
-            println(s"deleting... $dir")
-            dir.toPath.recursivelyDelete()
-          } catch {
-            case _: Exception =>
-          }
-      )
 }
